unconvert: drop redundant int64() on Timeval.Usec in io_uring bench
On Linux, syscall.Timeval.Usec is already int64 (the file is io_uring,
which compiles for Linux only), so int64(ru.Utime.Usec) and
int64(ru.Stime.Usec) are no-ops.

+++ pkg/io_uring/bench_test.go
@@ -31,8 +31,8 @@ func snapshotRusage(b *testing.B) rusageDelta {
 		b.Fatalf("Getrusage: %v", err)
 	}
 	return rusageDelta{
-		user: ru.Utime.Sec*1e6 + int64(ru.Utime.Usec),
-		sys:  ru.Stime.Sec*1e6 + int64(ru.Stime.Usec),
+		user: ru.Utime.Sec*1e6 + ru.Utime.Usec,
+		sys:  ru.Stime.Sec*1e6 + ru.Stime.Usec,
 		maj:  ru.Majflt,
 		nvcs: ru.Nvcsw,
 		nivs: ru.Nivcsw,
